renamed to cytoscape-init
